stop returning meaningful JournalPosition results from subscribeStreamTemporary
Summary:
Watchman, the only consumer of subscribeStreamTemporary today, does
not use the published JournalPosition values. Instead, it waits a bit,
and calls either getCurrentJournalPosition or
getFilesChangedSince. This is a good thing, because write traffic to a
filesystem can easily exceed the rate at which Watchman (or even
streaming Thrift!) can consume it, spiking memory usage and degrading
performance.

The next diff in this stack will take advantage of this clarified
communication protocol to reduce edenfs memory usage under heavy write
traffic.

Reviewed By: wez

Differential Revision: D24089811

fbshipit-source-id: 49de6be24a78934888162a577a554d50e2fff355