ui: disable ui.threaded by default
Summary:
Even if ui.threaded is off, we're seeing sys call interrupt issues. I
think this is because ui.threaded defaults to true, and we check it's value
before loading the repo config (which includes the dynamicconfig to turn it
off). It's already off in dynamicconfig, so let's turn it off by default.

This highlights an issue with dynamicconfig, in that it can't deploy pre-repo
configs at the moment.

As for the actual interrupted syscall, it looks to be caused by SIGCHLD signals
from child processes we spawn (possibly the dynamicconfig child process).

Reviewed By: quark-zju

Differential Revision: D24401521

fbshipit-source-id: 0372ffc7297fa6fe6a7e4f0ae69299b252c3632e