mononoke: add --overwrite option to rsync
Summary:
Previously `mononoke_admin rsync` didn't overwrite files i.e. if a target
directory has a file with the same name as in source directory then it won't be overwritten.

This diff adds an option to make it possible to overwrite a file if a target
directory has a file with the same name but its content is different. However
note that if a file has to be overwritten then target file is going to be removed in one
commit and then copied in the second (i.e. we'll create two commits instead of
one). The reason for doing that is to preserve the history of the original file
(i.e. history from `from-directory`).

Reviewed By: aslpavel

Differential Revision: D24538199

fbshipit-source-id: 792162c4e5ad81fb6949dd95eb1322696ec011ea